@@ -6,7 +6,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -14,7 +14,7 @@ dependencies:
   
   flutter_state_notifier: ^0.7.0
   flutter_riverpod: ^0.14.0
-  freezed_annotation:
+  freezed_annotation: ^0.14.2
   webview_flutter: ^2.0.0
   chopper: ^4.0.0
 
@@ -24,9 +24,11 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  freezed:
+  freezed: ^0.14.2
   build_runner: ^2.0.6
   chopper_generator: ^4.0.0
+  json_serializable: ^4.1.3
+  pedantic_mono:
 
 flutter:
 
