@@ -1,10 +1,17 @@
-// user.dart
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:flutter/foundation.dart';
 
 part 'article.freezed.dart';
+part 'article.g.dart';
 
 @freezed
 abstract class Article with _$Article {
- const factory Article({@Default('') String name, @Default(0) int age}) = Article;
+
+  const factory Article({
+    @Default('') @JsonKey(name: 'title') String? title,
+    @Default('') @JsonKey(name: 'userName') String? userName,
+    @Default('') @JsonKey(name: 'userIconUrl') String? userIconUrl,
+    @Default('') @JsonKey(name: 'url') String? url,
+  }) = _Article;
+
+  factory Article.fromJson(Map<String, dynamic> json) => _$ArticleFromJson(json);
 }
